requirements.txt
PyQt6~=6.5.1
httpx
pycryptodome~=3.18.0